Added Animation Class
The animation class handles animations with ease. To use it, a vec2 array of image coords must be created. The animate function then iterates through these vectors and crops the image according to the coords at the current index. The frequency of the iteration is determined by the delta time parameter in milliSeconds.

States and entities are for testing purposes.